refactor: presentation Layer を導入 (#6)
