+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/presentation/detail/DetailFragment.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.presentation.detail
 
 import android.os.Bundle
 import android.util.Log
@@ -9,7 +9,8 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import coil.load
-import jp.co.yumemi.android.code_check.MainActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.presentation.MainActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.R
 import jp.co.yumemi.android.code_check.databinding.FragmentDetailBinding
 import jp.co.yumemi.android.code_check.models.Repository
