+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/presentation/main/MainFragmentAdapter.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.presentation.main
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.R
 import jp.co.yumemi.android.code_check.models.Repository
 import jp.co.yumemi.android.code_check.util.diff_util
