+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/presentation/main/MainUiState.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.presentation.main
 
 import jp.co.yumemi.android.code_check.models.Repository
