+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/presentation/main/MainViewModel.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.presentation.main
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
