+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="jp.co.yumemi.android.code_check.MainFragment"
+        android:name="jp.co.yumemi.android.code_check.presentation.main.MainFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_main">
         <action
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/detailFragment"
-        android:name="jp.co.yumemi.android.code_check.DetailFragment"
+        android:name="jp.co.yumemi.android.code_check.presentation.detail.DetailFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_detail">
         <argument
